+++ docs/source/conf.py
@@ -134,8 +134,4 @@ intersphinx_mapping = {
     "python": ("https://docs.python.org/3", None),
     "numpy": ("https://numpy.org/doc/stable/", None),
     "scipy": ("https://docs.scipy.org/doc/scipy/", None),
-    "arepyextras.quality": (
-        "http://intranet.aresys.it/sardashboard/develop/arepyextras/arepyextras-quality/docs/latest/",
-        None,
-    ),
 }