+++ scripts/test_ptar_analysis.py
@@ -3,7 +3,7 @@
 
 """SCT: test PTAR analysis script"""
 
-from arepyextras.quality.target_ambiguity_ratio_analysis.graphical_output import ambiguities_graphs
+from perseo_quality.tar_analysis.graphical_output import ambiguities_graphs
 
 from sct.analyses.ambiguity_ratio_analysis import sct_point_target_ambiguity_ratio_analysis
 from sct.configuration.logger import ConsoleHandler, enable_quality_logger, sct_logger