+++ sct/io/point_target_manager.py
@@ -12,11 +12,11 @@ from pathlib import Path
 
 import numpy as np
 import pandas as pd
-from arepyextras.quality.core.signal_processing import convert_to_db
 from arepytools.geometry.conversions import llh2xyz, xyz2llh
 from arepytools.io import PointSetProduct, read_point_targets_file
 from arepytools.io.io_support import NominalPointTarget
 from arepytools.timing.precisedatetime import PreciseDateTime
+from perseo_quality.core.signal_processing import convert_to_db
 
 from sct import csv_template